third_party/WebKit/LayoutTests/http/tests/security/cross-frame-window-to-tostring-expected.txt

 
PASS successfullyParsed is true

TEST COMPLETE
PASS window + '' is "[object Window]"
PASS window.toString() is "[object Window]"
FAIL Object.prototype.toString.call(window) should be [object Window]. Was [object global].
FAIL Window.prototype + '' should be [object WindowPrototype]. Was [object Window].
FAIL Window.prototype.toString() should be [object WindowPrototype]. Was [object Window].
FAIL Object.prototype.toString.call(Window.prototype) should be [object WindowPrototype]. Was [object Object].
FAIL window + '' should be [object iii]. Was [object Window].
FAIL window.toString() should be [object iii]. Was [object Window].
PASS Object.prototype.toString.call(window) is "[object iii]"
FAIL Window.prototype + '' should be [object WindowPrototype]. Was [object Window].
FAIL Window.prototype.toString() should be [object WindowPrototype]. Was [object Window].
FAIL Object.prototype.toString.call(Window.prototype) should be [object WindowPrototype]. Was [object Object].
FAIL window + '' should be [object jjj]. Was [object Window].
FAIL window.toString() should be [object jjj]. Was [object Window].
PASS Object.prototype.toString.call(window) is "[object jjj]"
FAIL Window.prototype + '' should be [object jjj]. Was [object Window].
FAIL Window.prototype.toString() should be [object jjj]. Was [object Window].
PASS Object.prototype.toString.call(Window.prototype) is "[object jjj]"

Testing local iframe
FAIL w + '' should be [object inner]. Was [object Window].
FAIL w.toString() should be [object inner]. Was [object Window].
PASS Object.prototype.toString.call(w) is "[object inner]"
FAIL w + '' should be [object kkk]. Was [object Window].
FAIL w.toString() should be [object kkk]. Was [object Window].
PASS Object.prototype.toString.call(w) is "[object kkk]"
FAIL w + '' should be [object inner]. Was [object Window].
FAIL w.toString() should be [object inner]. Was [object Window].
FAIL Object.prototype.toString.call(w) should be [object inner]. Was [object global].

Testing remote iframe
PASS w + '' threw exception SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame..
FAIL w.toString() should throw SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.. Was [object Window].
PASS Object.prototype.toString.call(w) threw exception SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame..
